pyproject.toml
[project]
name = "manganize"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "ddgs>=9.9.1",
    "duckduckgo-search>=8.1.1",
    "google-genai>=1.52.0",
    "langchain>=1.1.0",
    "langchain-community>=0.4.1",
    "langchain-google-genai>=3.2.0",
    "langgraph>=1.0.4",
    "langgraph-cli[inmem]>=0.4.7",
    "markitdown[all]>=0.1.0",
    "pillow>=12.0.0",
    "playwright>=1.49.0",
    "pyyaml>=6.0.0",
    "tenacity>=9.1.2",
    # Web application dependencies
    "fastapi[standard]>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "jinja2>=3.1.0",
    "sse-starlette>=2.0.0",
    "python-multipart>=0.0.20",
    "aiosqlite>=0.20.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "alembic>=1.15.0",
    "pydantic-settings>=2.7.0",
    "slowapi>=0.1.9",
    "python-dotenv>=1.0.0",
]


[dependency-groups]
dev = ["ruff>=0.14.7", "ty>=0.0.7"]
